pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kipbot"
version = "0.1.0"
description = "Super User-Friendly Personal AI Assistant"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Team TNT Labs" }]
keywords = ["ai", "agent", "chatbot", "assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "python-telegram-bot>=20.0",
    "discord.py>=2.3.0",
    "croniter>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
kakao = [
    "flask>=3.0.0",
]

[project.scripts]
kipbot = "kipbot.cli.commands:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.hatch.build.targets.wheel]
packages = ["kipbot"]
